+++ doc/changelog.rst
@@ -38,6 +38,8 @@ Front
 Xtheme
 ~~~~~~
 
+- Add screenshot support for stylesheets
+
 Classic Gray Theme
 ~~~~~~~~~~~~~~~~~~